refactor: Entity 패키지 및 파일 이름 변경
